requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
Flask==3.0.0
gunicorn==21.2.0
requests==2.31.0